requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
torch==2.2.2
# torch==2.5.1
torchvision==0.17.2
pillow
pandas
python-multipart